add delegate to build Description objects without information loss
class TestMethodDescriptionBuilder is introduced as a builder for junit
Description objects. As those have no suited factory method, reflection is
used to access the private constructors. Implemented as an _enum_ all
the reflection stuff is lazy evaluated and executed only on purpose.
Fallback is kept with the initial implementation.

fix #103